Reorganization of test subsystem.
This rather large commit does effectively not touch the functionality
of the single probes, but reorganizes everything to a great
deal. Details are explained in the new `README.md`.

New features include better readability, easier extension of the
scripts, brief progress reporting on the console, a comprehensive test
report, and a detailed test protocol.

A number of minor bug fixes and typos have been applied.